requirements.txt
PySide6>=6.5.0
easyocr>=1.7.0
sentence-transformers>=2.2.2
pandas>=2.0.0
openpyxl>=3.1.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
torch>=2.0.0
PyMuPDF>=1.23.0
python-docx>=1.0.0
